Move ExoOfflineManager imp to ".exo" package
