+++ tvplayer/src/main/java/com/kaltura/tvplayer/offline/exo/WorkManagerScheduler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.kaltura.tvplayer.offline;
+package com.kaltura.tvplayer.offline.exo;
 
 import android.annotation.TargetApi;
 import android.content.Context;
